shared/requirements.txt
PyJWT==2.8.0
prometheus-client==0.19.0
fastapi==0.104.1
pydantic==2.5.0